ref: Extract "move endpoints" logic to a separate class.
